pyproject.toml
[project]
name = "aicubench"
version = "0.1.0"
description = "A benchmark tool for generative AI systems (ComfyUI, Whisper, SDXL, etc.)"
authors = [
    { name="AICU Inc.", email="support@aicu.ai" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "torch",
    "transformers",
    "psutil",
    "requests",
    "rich",
    "pyyaml",
    "pillow",
]

[project.scripts]
aicubench = "aicubench.main:main"
aicubench-compare = "aicubench.compare.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"